requirements.txt
streamlit
sentence-transformers
openai
chromadb
pysqlite3-binary
panda
